enhanced_trading_system/requirements.txt
python-binance==1.0.19
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.1
websocket-client>=1.2.1
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0